requirements.txt
streamlit==1.32.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.12.0
pandas==2.1.0
numpy==1.26.0
matplotlib==3.7.2
beautifulsoup4==4.12.2
langchain==0.1.0
langchain-community==0.0.13
pypdf==3.17.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pdfplumber==0.10.3
chromadb==0.4.18
tiktoken==0.5.1
